requirements.txt
datasets==3.0.1
ujson==5.10.0
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy>=1.26
